Fix scope issue for fork start and range

+++ src/main.cpp
@@ -1725,8 +1725,8 @@ namespace Consensus
 bool CheckTxInputs(const CTransaction& tx, CValidationState& state, const CCoinsViewCache& inputs, int nSpendHeight, const Consensus::Params& consensusParams)
 {   
     // const CChainParams& chainparams = Params();
-    int64_t nForkStartHeight = consensusParams.nForkStartHeight;
-    int64_t nForkHeightRange = consensusParams.nForkHeightRange;
+    int nForkStartHeight = consensusParams.nForkStartHeight;
+    int nForkHeightRange = consensusParams.nForkHeightRange;
     // This doesn't trigger the DoS code on purpose; if it did, it would make it easier
     // for an attacker to attempt to split the network.
     if (!inputs.HaveInputs(tx))
